Too many changes
Add vim-colors-solarized
Update agnostic to new master
Remove python shim and go with pyenv FTW!
Update gpg-agent.conf to remove things we are specifying on the command
line (probably I should just remove this file altogether).
Add a key for home ssh key to sshcontrol -- this needs to be synced to
the other machines.